Update activity_mood_log.xml
My main focus here is trying to get the current time and time to show up on the mood log page. I used the code from here: https://www.tutorialspoint.com/how-do-i-display-the-current-date-and-time-in-an-android-application
Since the mood log page isn't connected to the previous page, I can't see the result in the emulator. But I assume the date/time is not showing up.
Files I added to:
activity_mood_log.xml
MainActivity.java
MoodlogActivity.java
AndroidManifest.xml

-Kenzie